Update file in Spanish.
Latin-American Spanish with some corrections on style, some translations not previously translated, some important changes to clarify, especially "peer" which is "usuario que comparte" (user that shares) as we have no word like "peer" in Spanish (the literal translation of "peer" (pareja) is the equivalent in Spanish to "mate" (couple) so it makes no sense in this context) and "buffering" in english makes a lot of sense, but in Spanish we have to say "the delay in the loading of the buffer" (el retraso al cargar el búfer) as there is no equivalent to the construction "buffering", so I had to explain it. The rest is just style corrections. Thanks you so much for having Hola for the world. If you have any questions please e-mail me at alekstestut@gmail.com.